Cyanotype

Documentation for Cyanotype.

Cyanotype.BpBatchNorm

Cyanotype.BpChannelExpansionConv

Cyanotype.BpConv

Cyanotype.BpDConv

Cyanotype.BpDepthwiseConv

Cyanotype.BpGroupNorm

Cyanotype.BpHybridAtrouConv

Cyanotype.BpInstanceNorm

Cyanotype.BpMBConv

Cyanotype.BpNConv

Cyanotype.BpPointwiseConv

Cyanotype.BpSqueezeExcitation

Cyanotype.ConvUpSamplerBp

Cyanotype.KwargsMapping

Cyanotype.LinearUpSamplerBp

Cyanotype.MaxDownSamplerBp

Cyanotype.MeanDownSamplerBp

Cyanotype.NearestUpSamplerBp

Cyanotype.PixelClassifierBp

Cyanotype.chcat

Cyanotype.cyanotype

Cyanotype.make

Cyanotype.spread

Cyanotype.@cyanotype

Cyanotype.BpBatchNorm — Type

Wraps a Flux.Batchnorm

Keyword arguments:

activation: activation function (default identity)

initshift: see initβ (default zeros32)

initscale: see initγ (default ones32)

affine: see affine (default true)

trackstats: see track_stats (default true)

epsilon: see ϵ (default 1.0e-5)

momentum: see momentum (default 0.1)

Cyanotype.BpChannelExpansionConv — Type

Cyanotype.BpConv — Type

BpConv(; kwargs...)

A cyanotype blueprint describing a convolutionnal module or layer depending om the value of normalization argument.

Keyword arguments:

volume: indicates a building process for three-dimensionnal data (default false)

activation: activation function (default identity)

normalization:

depthwise:

revnorm:

preactivation:

bias:

stride: see stride (default 1)

pad: see pad (default Flux.SamePad())

dilation: see dilation (default 1)

groups: see groups (default 1)

init: see init (default glorot_uniform)

Cyanotype.BpDConv — Type

BpDConv(; kwargs)

Describes a convolutionnal module formed by two successive convolutionnal modules.

Keyword arguments:

volume: indicates a building process for three-dimensionnal data (default false)

conv1 is not documented

conv2 is not documented

Cyanotype.BpDepthwiseConv — Type

Cyanotype.BpGroupNorm — Type

BpGroupNorm(; kwargs...)

Describes a building process for a Groupnorm layer. make(channels, bp::CyGroupNorm)

Keyword arguments:

activation: activation function (default identity)

groups: the number of groups passed to GroupNorm

constructor

initshift: see initβ (default zeros32)

initscale: see initγ (default ones32)

affine: see affine (default true)

trackstats: see track_stats (default false)

epsilon: see ϵ (default 1.0e-5)

momentum: see momentum (default 0.1)

Cyanotype.BpHybridAtrouConv — Type

aka Hybrid Dilated Convolution paper example example

Keyword arguments:

dilation_rates is not documented

conv is not documented

Cyanotype.BpInstanceNorm — Type

BpInstanceNorm(; kwargs...)

Describes a building process for a InstanceNorm layer. make(channels, bp::CyInstanceNorm)

Keyword arguments:

activation: activation function (default identity)

initshift: see initβ (default zeros32)

initscale: see initγ (default ones32)

affine: see affine (default false)

trackstats: see track_stats (default false)

epsilon: see ϵ (default 1.0e-5)

momentum: see momentum (default 0.1)

Cyanotype.BpMBConv — Type

Cyanotype.BpNConv — Type

Template describing a module with N NConvBp repeated.

Keyword arguments:

convolution is not documented

nrepeat is not documented

Cyanotype.BpPointwiseConv — Type

Cyanotype.BpSqueezeExcitation — Type

https://arxiv.org/pdf/1709.01507.pdf

Cyanotype.ConvUpSamplerBp — Type

Keyword arguments:

volume: indicates a building process for three-dimensionnal data (default false)

scale is not documented

Cyanotype.KwargsMapping — Type

KwargsMapping(; flux_function = :notflux, field_names = (), flargs = (),
                ftypes = (), defval = ())

Define a mapping of keyword arguments to interface a blueprint with a Flux function or constructor.

Cyanotype.LinearUpSamplerBp — Type

Keyword arguments:

volume: indicates a building process for three-dimensionnal data (default false)

scale is not documented

Cyanotype.MaxDownSamplerBp — Type

Keyword arguments:

volume: indicates a building process for three-dimensionnal data (default false)

wsize is not documented

Cyanotype.MeanDownSamplerBp — Type

Keyword arguments:

volume: indicates a building process for three-dimensionnal data (default false)

wsize is not documented

Cyanotype.NearestUpSamplerBp — Type

Keyword arguments:

volume: indicates a building process for three-dimensionnal data (default false)

scale is not documented

Cyanotype.PixelClassifierBp — Type

Keyword arguments:

volume: indicates a building process for three-dimensionnal data (default false)

nclasses is not documented

Cyanotype.chcat — Method

chcat(x...)

Concatenate the image data along the dimension corresponding to the channels. Image data should be stored in WHCN order (width, height, channels, batch) or WHDCN (width, height, depth, channels, batch) in 3D context. Channels are assumed to be the penultimate dimension.

Example

julia> x1 = rand(32, 32, 4, 6); # a batch of 6 images (32x32) with 4 channels

julia> x2 = rand(32, 32, 4, 6); # a batch of 6 images (32x32) with 4 channels

julia> chcat(x1, x2) |> size
(32, 32, 8, 6)

Cyanotype.cyanotype — Function

cyanotype(bp::AbstractBlueprint; kwargs...)

Create a new blueprint from bp with the modifications defined by kwargs. This method is automatically generated by the @cyanotype macro during the process of defining a blueprint.

Cyanotype.make — Function

Cyanotype.spread — Method

spread(bp; kwargs...)

Return a new blueprint with kwargs spreaded over all fields of bp. This function allows to modify all nested cyanotypes at once.

# Nice printing of large objects
using GarishPrint

# Create a blueprint for a Hybrid A-trou Convolution module
hac = CyHybridAtrouConv();
# By default activation functions are relu
pprint(hac)

# Create a blueprint for a double convolution module with the second convolution as a usual
# convolutionnal layer
conv2 = CyDoubleConv(; convolution1 = hac, convolution2 = CyConv());
# By default activation functions are relu
pprint(conv2)

# Now let change all activation functions from relu to leakyrelu
spread(conv2; activation = leakyrelu) |> pprint

Cyanotype.@cyanotype — Macro

@cyanotype(expr)
@cyanotype begin
    [kmap]
    [doc]
    expr
end

Defines a blueprint DataType with documentation doc and a struct declaration defined in expr. If the blueprint directly refers to a Flux function or constructor, kmap is the name of keyword arguments mapping. If there is some fields documentation in expr, it is automatically appended to doc.

Automatically generated functions:

FooBluePrint(; kwargs...): keyword argument constructor for FooBluePrint

mapping(::FooBluePrint): return, if defined, the mapping kmap

cyanotype(bp::FooBluePrint; kwargs...)

kwargs(bp::FooBluePrint): return, if defined, a Dict with the keyword arguments for

the Flux function or constructor, it can be used as follow: flux_function(arg1, arg2; kwargs(bp)...)

Example:

using Cyanotype: @cyanotype

@cyanotype begin
    """
    A FooBlueprint as example.
    """
    struct FooBlueprint{A<:Function}
        """`activation`: activation function"""
        activation::A = relu
    end
end

For the keyword arguments mapping usage, see KwargsMapping documentation.