Cyanotype

Documentation for Cyanotype.

Cyanotype.AxialDWConvBp

Cyanotype.BatchNormBp

Cyanotype.CBAMBp

Cyanotype.ChannelAttentionBp

Cyanotype.ChannelExpansionConvBp

Cyanotype.ConvBp

Cyanotype.ConvTransposeUpsamplerBp

Cyanotype.DepthwiseConvBp

Cyanotype.DoubleConvBp

Cyanotype.EfficientNetBp

Cyanotype.EfficientNetStageBp

Cyanotype.EfficientUNetBp

Cyanotype.FusedMbConvBp

Cyanotype.GroupNormBp

Cyanotype.HybridAtrouConvBp

Cyanotype.InstanceNormBp

Cyanotype.KwargsMapping

Cyanotype.LabelClassifierBp

Cyanotype.LinearUpsamplerBp

Cyanotype.MaxDownsamplerBp

Cyanotype.MbConvBp

Cyanotype.MeanDownsamplerBp

Cyanotype.NConvBp

Cyanotype.NearestUpsamplerBp

Cyanotype.PixelClassifierBp

Cyanotype.PixelMapBp

Cyanotype.PixelShuffleUpsamplerBp

Cyanotype.PointwiseConvBp

Cyanotype.ResCBAMBp

Cyanotype.ResidualConvBp

Cyanotype.SpatialAttentionBp

Cyanotype.SqueezeExcitationBp

Cyanotype.UBridgeBp

Cyanotype.UDecoderBp

Cyanotype.UEncoderBp

Cyanotype.UNetBp

Cyanotype.chcat

Cyanotype.cyanotype

Cyanotype.make

Cyanotype.spread

Cyanotype.uchain

Cyanotype.@cyanotype

Cyanotype.AxialDWConvBp — Type

https://arxiv.org/pdf/2306.16103v2.pdf

Keyword arguments:

volume: indicates a building process for three-dimensionnal data (default false)

activation: activation function (default identity)

stride is not documented

pad is not documented

dilation is not documented

init is not documented

normalization is not documented

Cyanotype.BatchNormBp — Type

Wraps a Flux.Batchnorm

Keyword arguments:

activation: activation function (default identity)

initshift: see initβ (default zeros32)

initscale: see initγ (default ones32)

affine: see affine (default true)

trackstats: see track_stats (default true)

epsilon: see eps (default 1.0e-5)

momentum: see momentum (default 0.1)

Cyanotype.CBAMBp — Type

Cyanotype.ChannelAttentionBp — Type

Cyanotype.ChannelExpansionConvBp — Type

Cyanotype.ConvBp — Type

ConvBp(; kwargs...)

A cyanotype blueprint describing a convolutionnal module or layer depending om the value of normalization argument.

Keyword arguments:

volume: indicates a building process for three-dimensionnal data (default false)

activation: activation function (default identity)

normalization:

depthwise:

revnorm:

preactivation:

bias:

stride: see stride (default 1)

pad: see pad (default Flux.SamePad())

dilation: see dilation (default 1)

groups: see groups (default 1)

init: see init (default glorot_uniform)

Cyanotype.ConvTransposeUpsamplerBp — Type

Keyword arguments:

volume: indicates a building process for three-dimensionnal data (default false)

scale is not documented

bias is not documented

init is not documented

Cyanotype.DepthwiseConvBp — Type

Cyanotype.DoubleConvBp — Type

DoubleConvBp(; kwargs)

Describes a convolutionnal module formed by two successive convolutionnal modules.

Keyword arguments:

conv1 is not documented

conv2 is not documented

Cyanotype.EfficientNetBp — Type

Cyanotype.EfficientNetStageBp — Type

Keyword arguments:

ksize is not documented

outchannels is not documented

nrepeat is not documented

convolution is not documented

widthscaling is not documented

depthscaling is not documented

Cyanotype.EfficientUNetBp — Type

Keyword arguments:

backbone is not documented

decoder is not documented

head is not documented

Cyanotype.FusedMbConvBp — Type

Cyanotype.GroupNormBp — Type

GroupNormBp(; kwargs...)

Describes a building process for a Groupnorm layer. make(channels, bp::CyGroupNorm)

Keyword arguments:

activation: activation function (default identity)

groups: the number of groups passed to GroupNorm

constructor

initshift: see initβ (default zeros32)

initscale: see initγ (default ones32)

affine: see affine (default true)

trackstats: see track_stats (default false)

epsilon: see eps (default 1.0e-5)

momentum: see momentum (default 0.1)

Cyanotype.HybridAtrouConvBp — Type

aka Hybrid Dilated Convolution paper example example

Cyanotype.InstanceNormBp — Type

InstanceNormBp(; kwargs...)

Describes a building process for a InstanceNorm layer. make(channels, bp::CyInstanceNorm)

Keyword arguments:

activation: activation function (default identity)

initshift: see initβ (default zeros32)

initscale: see initγ (default ones32)

affine: see affine (default false)

trackstats: see track_stats (default false)

epsilon: see eps (default 1.0e-5)

momentum: see momentum (default 0.1)

Cyanotype.KwargsMapping — Type

KwargsMapping(; flux_function = :notflux, field_names = (), flargs = (),
                ftypes = (), defval = ())

Define a mapping of keyword arguments to interface a blueprint with a Flux function or constructor.

Cyanotype.LabelClassifierBp — Type

Keyword arguments:

nclasses is not documented

dropout is not documented

Cyanotype.LinearUpsamplerBp — Type

Keyword arguments:

volume: indicates a building process for three-dimensionnal data (default false)

scale is not documented

Cyanotype.MaxDownsamplerBp — Type

Keyword arguments:

volume: indicates a building process for three-dimensionnal data (default false)

wsize is not documented

Cyanotype.MbConvBp — Type

Cyanotype.MeanDownsamplerBp — Type

Keyword arguments:

volume: indicates a building process for three-dimensionnal data (default false)

wsize is not documented

Cyanotype.NConvBp — Type

Template describing a module with N NConvBp repeated.

Keyword arguments:

convolutions is not documented

Cyanotype.NearestUpsamplerBp — Type

Keyword arguments:

volume: indicates a building process for three-dimensionnal data (default false)

scale is not documented

Cyanotype.PixelClassifierBp — Type

Cyanotype.PixelMapBp — Type

Cyanotype.PixelShuffleUpsamplerBp — Type

Cyanotype.PointwiseConvBp — Type

Cyanotype.ResCBAMBp — Type

Cyanotype.ResidualConvBp — Type

Keyword arguments:

convolution is not documented

connector is not documented

Cyanotype.SpatialAttentionBp — Type

Cyanotype.SqueezeExcitationBp — Type

https://arxiv.org/pdf/1709.01507.pdf

Cyanotype.UBridgeBp — Type

Keyword arguments:

convolution is not documented

downsampler is not documented

upsampler is not documented

Cyanotype.UDecoderBp — Type

Keyword arguments:

convolution is not documented

upsampler is not documented

Cyanotype.UEncoderBp — Type

Keyword arguments:

convolution is not documented

downsampler is not documented

Cyanotype.UNetBp — Type

Keyword arguments:

inchannels is not documented

nlevels is not documented

basewidth is not documented

expansion is not documented

ksize is not documented

encoder is not documented

decoder is not documented

bridge is not documented

stem is not documented

path is not documented

head is not documented

top is not documented

Cyanotype.chcat — Method

chcat(x...)

Concatenate the image data along the dimension corresponding to the channels. Image data should be stored in WHCN order (width, height, channels, batch) or WHDCN (width, height, depth, channels, batch) in 3D context. Channels are assumed to be the penultimate dimension.

Example

julia> x1 = rand(32, 32, 4, 6); # a batch of 6 images (32x32) with 4 channels

julia> x2 = rand(32, 32, 4, 6); # a batch of 6 images (32x32) with 4 channels

julia> chcat(x1, x2) |> size
(32, 32, 8, 6)

Cyanotype.cyanotype — Function

cyanotype(bp::AbstractBlueprint; kwargs...)

Create a new blueprint from bp with the modifications defined by kwargs. This method is automatically generated by the @cyanotype macro during the process of defining a blueprint.

Cyanotype.make — Function

Cyanotype.spread — Method

spread(bp; kwargs...)

Return a new blueprint with kwargs spreaded over all fields of bp. This function allows to modify all nested cyanotypes at once.

# Nice printing of large objects
using GarishPrint

# Create a blueprint for a Hybrid A-trou Convolution module
hac = HybridAtrouConvBp();
# By default activation functions are relu
pprint(hac)

# Create a blueprint for a double convolution module with the second convolution as a usual
# convolutionnal layer
conv2 = DoubleConvBp(; convolution1 = hac, convolution2 = ConvBp());
# By default activation functions are relu
pprint(conv2)

# Now let change all activation functions from relu to leakyrelu
spread(conv2; activation = leakyrelu) |> pprint

Cyanotype.uchain — Method

uchain(; encoders, decoders, bridge, connection, [paths])

Build a Chain with U-Net like architecture. encoders and decoders are vectors of encoding/decoding blocks, from top to bottom (see diagram below). bridge is the bottom part of U-Net architecture. Each level of the U-Net is connected through a channel concatenation (◊ symbol in the diagram below).

Notes :

usually encoder unit starts with a 'MaxPool' to downsample image by 2, except the first

level encoder.

usually decoder unit ends with a 'ConvTranspose' to upsample image by 2, except the first

level decoder.

┌─────────┐                                                          ┌─────────┐
│Encoder 1│                                                          │Decoder 1│
└────┬────┘                                                          └─────────┘
     │                                                                    ▴
     │                                                                    │
     ├───────────────────────────────────────────────────────────────────▸◊
     │    ┌─────────┐                                     ┌─────────┐     │
     └───▸│Encoder 2│                                     │Decoder 2├─────┘
          └────┬────┘                                     └─────────┘
               │                                               ▴
               │                                               │
               ├──────────────────────────────────────────────▸◊
               │     ┌─────────┐               ┌─────────┐     │
               └────▸│Encoder 3│               │Decoder 3├─────┘
                     └────┬────┘               └─────────┘
                          │                         ▴
                          │                         │
                          ├────────────────────────▸◊
                          │       ┌─────────┐       │
                          └──────▸│ Bridge  ├───────┘
                                  └─────────┘

See also chcat.

Cyanotype.@cyanotype — Macro

@cyanotype(expr)
@cyanotype begin
    [kmap]
    [doc]
    expr
end

Defines a blueprint DataType with documentation doc and a struct declaration defined in expr. If the blueprint directly refers to a Flux function or constructor, kmap is the name of keyword arguments mapping. If there is some fields documentation in expr, it is automatically appended to doc.

Automatically generated functions:

FooBluePrint(; kwargs...): keyword argument constructor for FooBluePrint

mapping(::FooBluePrint): return, if defined, the mapping kmap

cyanotype(bp::FooBluePrint; kwargs...)

kwargs(bp::FooBluePrint): return, if defined, a Dict with the keyword arguments for

the Flux function or constructor, it can be used as follow: flux_function(arg1, arg2; kwargs(bp)...)

Example:

using Cyanotype: @cyanotype

@cyanotype begin
    """
    A FooBlueprint as example.
    """
    struct FooBlueprint{A<:Function}
        """`activation`: activation function"""
        activation::A = relu
    end
end

For the keyword arguments mapping usage, see KwargsMapping documentation.