Cyanotype

Documentation for Cyanotype.

Cyanotype.AxialDWConvBp

Cyanotype.BatchNormBp

Cyanotype.CBAMBp

Cyanotype.ChannelAttentionBp

Cyanotype.ChannelExpansionConvBp

Cyanotype.ConvBp

Cyanotype.ConvTransposeUpsamplerBp

Cyanotype.DepthwiseConvBp

Cyanotype.DoubleConvBp

Cyanotype.EfficientNetBp

Cyanotype.EfficientNetStageBp

Cyanotype.EfficientUNetBp

Cyanotype.FusedMbConvBp

Cyanotype.GroupNormBp

Cyanotype.HybridAtrouConvBp

Cyanotype.InstanceNormBp

Cyanotype.KwargsMapping

Cyanotype.LabelClassifierBp

Cyanotype.LinearUpsamplerBp

Cyanotype.MaxDownsamplerBp

Cyanotype.MbConvBp

Cyanotype.MeanDownsamplerBp

Cyanotype.MeanMaxDownsamplerBp

Cyanotype.NConvBp

Cyanotype.NearestUpsamplerBp

Cyanotype.PixelClassifierBp

Cyanotype.PixelMapBp

Cyanotype.PixelShuffleUpsamplerBp

Cyanotype.PointwiseConvBp

Cyanotype.ResCBAMBp

Cyanotype.ResidualConvBp

Cyanotype.SpatialAttentionBp

Cyanotype.SqueezeExcitationBp

Cyanotype.UBridgeBp

Cyanotype.UDecoderBp

Cyanotype.UEncoderBp

Cyanotype.UNet2Bp

Cyanotype.UNetBp

Cyanotype.chcat

Cyanotype.chmaxpool

Cyanotype.chmeanpool

Cyanotype.chsoftmax

Cyanotype.cyanotype

Cyanotype.flatten_layers

Cyanotype.genk

Cyanotype.make

Cyanotype.spread

Cyanotype.spread

Cyanotype.uchain

Cyanotype.@activation

Cyanotype.@cyanotype

Cyanotype.@volume

Cyanotype.AxialDWConvBp — Type

https://arxiv.org/pdf/2306.16103v2.pdf

Keyword arguments:

vol: indicates a building process for three-dimensionnal data (default

false)

act: activation function (default identity)

stride is not documented

pad is not documented

dila is not documented

init is not documented

norm is not documented

skip is not documented

Cyanotype.BatchNormBp — Type

Wraps a Flux.Batchnorm

Keyword arguments:

act: activation function (default identity)

initshift: see initβ (default zeros32)

initscale: see initγ (default ones32)

affine: see affine (default true)

trackstats: see track_stats (default true)

epsilon: see eps (default 1.0e-5)

momentum: see momentum (default 0.1)

Cyanotype.CBAMBp — Type

Cyanotype.ChannelAttentionBp — Type

Cyanotype.ChannelExpansionConvBp — Type

Cyanotype.ConvBp — Type

ConvBp(; kwargs...)

A cyanotype blueprint describing a convolutionnal module or layer depending om the value of norm argument.

Keyword arguments:

vol: indicates a building process for three-dimensionnal data (default

false)

act: activation function (default identity)

norm:

dwise:

rvnorm:

preact:

bias:

stride: see stride (default 1)

pad: see pad (default Flux.SamePad())

dila: see dilation (default 1)

groups: see groups (default 1)

init: see init (default glorot_uniform)

Cyanotype.ConvTransposeUpsamplerBp — Type

Keyword arguments:

vol: indicates a building process for three-dimensionnal data (default

false)

scale is not documented

bias is not documented

init is not documented

Cyanotype.DepthwiseConvBp — Type

Cyanotype.DoubleConvBp — Type

DoubleConvBp(; kwargs)

Describes a convolutionnal module formed by two successive convolutionnal modules.

Keyword arguments:

conv2 is not documented

Cyanotype.EfficientNetBp — Type

Cyanotype.EfficientNetStageBp — Type

Keyword arguments:

wscaling is not documented

dscaling is not documented

Cyanotype.EfficientUNetBp — Type

Keyword arguments:

bbone is not documented

dec is not documented

head is not documented

Cyanotype.FusedMbConvBp — Type

Cyanotype.GroupNormBp — Type

GroupNormBp(; kwargs...)

Describes a building process for a Groupnorm layer. make(channels, bp::CyGroupNorm)

Keyword arguments:

act: activation function (default identity)

groups: the number of groups passed to GroupNorm

constructor

initshift: see initβ (default zeros32)

initscale: see initγ (default ones32)

affine: see affine (default true)

epsilon: see eps (default 1.0e-5)

momentum: see momentum (default 0.1)

Cyanotype.HybridAtrouConvBp — Type

aka Hybrid Dilated Convolution paper example example

Cyanotype.InstanceNormBp — Type

InstanceNormBp(; kwargs...)

Describes a building process for a InstanceNorm layer. make(channels, bp::CyInstanceNorm)

Keyword arguments:

act: activation function (default identity)

initshift: see initβ (default zeros32)

initscale: see initγ (default ones32)

affine: see affine (default false)

trackstats: see track_stats (default false)

epsilon: see eps (default 1.0e-5)

momentum: see momentum (default 0.1)

Cyanotype.KwargsMapping — Type

KwargsMapping(; flux_function = :notflux, field_names = (), flargs = (),
                ftypes = (), defval = ())

Defines a mapping of keyword arguments to interface a blueprint with a Flux function or constructor.

Cyanotype.LabelClassifierBp — Type

Keyword arguments:

dropout is not documented

Cyanotype.LinearUpsamplerBp — Type

Keyword arguments:

vol: indicates a building process for three-dimensionnal data (default

false)

scale is not documented

Cyanotype.MaxDownsamplerBp — Type

Keyword arguments:

vol: indicates a building process for three-dimensionnal data (default

false)

wsize is not documented

Cyanotype.MbConvBp — Type

Cyanotype.MeanDownsamplerBp — Type

Keyword arguments:

vol: indicates a building process for three-dimensionnal data (default

false)

wsize is not documented

Cyanotype.MeanMaxDownsamplerBp — Type

Cyanotype.NConvBp — Type

Template describing a module with N NConvBp repeated.

Keyword arguments:

Cyanotype.NearestUpsamplerBp — Type

Keyword arguments:

vol: indicates a building process for three-dimensionnal data (default

false)

scale is not documented

Cyanotype.PixelClassifierBp — Type

Cyanotype.PixelMapBp — Type

Cyanotype.PixelShuffleUpsamplerBp — Type

Cyanotype.PointwiseConvBp — Type

Cyanotype.ResCBAMBp — Type

Cyanotype.ResidualConvBp — Type

Keyword arguments:

connector is not documented

Cyanotype.SpatialAttentionBp — Type

Cyanotype.SqueezeExcitationBp — Type

https://arxiv.org/pdf/1709.01507.pdf

Cyanotype.UBridgeBp — Type

Keyword arguments:

Cyanotype.UDecoderBp — Type

Keyword arguments:

Cyanotype.UEncoderBp — Type

Keyword arguments:

Cyanotype.UNet2Bp — Type

Keyword arguments:

in_chs is not documented

nlvl is not documented

bwidth is not documented

expn is not documented

ksize is not documented

stem is not documented

path is not documented

head is not documented

top is not documented

residual is not documented

Cyanotype.UNetBp — Type

Keyword arguments:

in_chs is not documented

nlvl is not documented

bwidth is not documented

expn is not documented

ksize is not documented

stem is not documented

path is not documented

head is not documented

top is not documented

Cyanotype.chcat — Method

chcat(x...)

Concatenates the provided data along the channel dimension positioned as the penultimate dimension.

Example

julia> x1 = rand(32, 32, 4, 6); # a batch of 6 images (32x32) with 4 channels

julia> x2 = rand(32, 32, 4, 6); # a batch of 6 images (32x32) with 4 channels

julia> chcat(x1, x2) |> size
(32, 32, 8, 6)

Cyanotype.chmaxpool — Method

chmaxpool(x)

Pools all channels using the maximum function.

Example

julia> x = rand(2, 2, 3, 1);

julia> Cyanotype.chmaxpool(x)
2×2×1×1 Array{Float64, 4}:
[:, :, 1, 1] =
 0.930081  0.648555
 0.759567  0.803717

Cyanotype.chmeanpool — Method

chmeanpool(x)

Pools all channels using the mean function.

Example

```julia julia> x = rand(2, 2, 3, 1);

julia> chmeanpool(x) 2×2×1×1 Array{Float64, 4}: [:, :, 1, 1] = 0.64925 0.242134 0.153542 0.312039 ```

Cyanotype.chsoftmax — Method

chsoftmax(x)

Apply a softmax function along the channel dimension. This function is useful for pixel classification (semantic segmentation).

Example

julia> x = rand(2, 2, 3, 1);

julia> sum(chsoftmax(x), dims=3)
2×2×1×1 Array{Float64, 4}:
[:, :, 1, 1] =
 1.0  1.0
 1.0  1.0

See also PixelClassifierBp.

Cyanotype.cyanotype — Function

cyanotype(bp::AbstractBlueprint; kwargs...)

Creates a new blueprint from bp with the modifications defined by kwargs. This method is automatically generated by the @cyanotype macro during the process of defining a blueprint.

Cyanotype.flatten_layers — Method

flatten_layers(layers...)

Flatten a nested Vector/Tuple/Chain into a single Vector. If any layers contains identity function it is skipped.

Example

julia> flatten_layers(["a",("b","c",["d","e",identity])])
5-element Vector{Any}:
 "a"
 "b"
 "c"
 "d"
 "e"

Cyanotype.genk — Method

genk(k, vol)

Helping function used to generate a kernel tuple (k,k) or (k,k,k) if vol is true.

Cyanotype.make — Function

Cyanotype.spread — Method

spread(bp, fieldname, old => new)

Create a new blueprint from bp for the given fieldname, and replace each occurence of old with new. This function allows to modify all nested blueprints at once.

Example

julia> # Create a blueprint for a Hybrid A-trou Convolution unit. By default activation
       # functions are identity.

julia> hac = HybridAtrouConvBp();

julia> hac.conv.act |> println
identity

julia> # Create a blueprint for a double convolution unit with a second convolution as a
       # usual convolutionnal layer with `relu` as activation.

julia> conv = DoubleConvBp(; conv1 = hac, conv2 = ConvBp(act = relu));

julia> # Now let change all activation functions from relu to leakyrelu

julia> bp = spread(conv, :act, identity => leakyrelu);

julia> bp.conv1.conv.act |> println
leakyrelu

julia> bp.conv2.act |> println
relu

Cyanotype.spread — Method

spread(bp; kwargs...)

Create a new blueprint with kwargs spreaded over all fields of bp. This function allows to modify all nested blueprint at once.

Example

julia> # Create a blueprint for a Hybrid A-trou Convolution unit. By default activation
       # functions are `identity`.

julia> hac = HybridAtrouConvBp();

julia> hac.conv.act |> println
identity

julia> # Create a blueprint for a double convolution unit a the second convolution as a
       # usual convolutionnal layer. By default activation function is `identity`.

julia> conv = DoubleConvBp(; conv1 = hac, conv2 = ConvBp());

julia> # Now let change all activation functions from relu to leakyrelu

julia> bp = spread(conv; act = leakyrelu);

julia> bp.conv1.conv.act |> println
leakyrelu

julia> bp.conv2.act |> println
leakyrelu

Cyanotype.uchain — Method

uchain(; encoders, decoders, bridge, connection, [paths])

Build a Chain with U-Net like architecture. encoders and decoders are vectors of encoding/decoding blocks, from top to bottom (see diagram below). bridge is the bottom part of U-Net architecture. Each level of the U-Net is connected through a channel concatenation (◊ symbol in the diagram below).

Notes :

usually encoder unit starts with a 'MaxPool' to downsample image by 2, except the first

level encoder.

usually decoder unit ends with a 'ConvTranspose' to upsample image by 2, except the first

level decoder.

┌─────────┐                                                          ┌─────────┐
│Encoder 1│                                                          │Decoder 1│
└────┬────┘                                                          └─────────┘
     │                                                                    ▴
     │                                                                    │
     ├───────────────────────────────────────────────────────────────────▸◊
     │    ┌─────────┐                                     ┌─────────┐     │
     └───▸│Encoder 2│                                     │Decoder 2├─────┘
          └────┬────┘                                     └─────────┘
               │                                               ▴
               │                                               │
               ├──────────────────────────────────────────────▸◊
               │     ┌─────────┐               ┌─────────┐     │
               └────▸│Encoder 3│               │Decoder 3├─────┘
                     └────┬────┘               └─────────┘
                          │                         ▴
                          │                         │
                          ├────────────────────────▸◊
                          │       ┌─────────┐       │
                          └──────▸│ Bridge  ├───────┘
                                  └─────────┘

See also chcat.

Cyanotype.@activation — Macro

@activation(func)

Helping macro used to define an activation function within a blueprint definition.

Cyanotype.@cyanotype — Macro

@cyanotype(expr)
@cyanotype begin
    [kmap]
    [doc]
    expr
end

Defines a blueprint DataType with documentation doc and a struct declaration defined in expr. If the blueprint directly refers to a Flux function or constructor, kmap is the name of keyword arguments mapping. If there is some fields documentation in expr, it is automatically appended to doc.

Automatically generated functions:

FooBluePrint(; kwargs...): keyword argument constructor for FooBluePrint

mapping(::FooBluePrint): return, if defined, the mapping kmap

cyanotype(bp::FooBluePrint; kwargs...)

kwargs(bp::FooBluePrint): return, if defined, a Dict with the keyword arguments for

the Flux function or constructor, it can be used as follow: flux_function(arg1, arg2; kwargs(bp)...)

Example:

using Cyanotype

@cyanotype begin
    """
    A FooBlueprint as example.
    """
    struct FooBlueprint{A<:Function}
        """`act`: activation function"""
        act::A = relu
    end
end

For the keyword arguments mapping usage, see KwargsMapping documentation.

Cyanotype.@volume — Macro

@volume

Helping macro to define 'vol' field within a blueprint definition.