Cyanotype

Documentation for Cyanotype.

Cyanotype.BatchNormBp

Cyanotype.ConvBp

Cyanotype.DoubleConvBp

Cyanotype.GroupNormBp

Cyanotype.HybridAtrouConvBp

Cyanotype.InstanceNormBp

Cyanotype.KwargsMapping

Cyanotype.NConvBp

Cyanotype.NoNormBp

Cyanotype.SqueezeExciteBp

Cyanotype.cyanotype

Cyanotype.make

Cyanotype.spread

Cyanotype.@cyanotype

Cyanotype.BatchNormBp — Type

Wraps a Flux.Batchnorm

Keyword arguments:

activation: activation function (default relu)

init_shift: see initβ (default zeros32)

init_scale: see initγ (default ones32)

affine: see affine (default true)

track_stats: see track_stats (default true)

epsilon: see ϵ (default 1.0e-5)

momentum: see momentum (default 0.1)

Cyanotype.ConvBp — Type

ConvBp(; kwargs)

A cyanotype blueprint describing a convolutionnal module or layer depending om the value of normalization argument.

Keyword arguments:

volumetric: indicates a building process for three-dimensionnal data (default false)

normalization:

reverse_norm:

pre_activation:

use_bias:

init: see init (default glorot_uniform)

pad: see pad (default Flux.SamePad())

dilation: see dilation (default 1)

groups: see groups (default 1)

Cyanotype.DoubleConvBp — Type

DoubleConvBp(; kwargs)

Describes a convolutionnal module formed by two successive convolutionnal modules.

Keyword arguments:

volumetric: indicates a building process for three-dimensionnal data (default false)

convolution1 is not documented

convolution2 is not documented

Cyanotype.GroupNormBp — Type

GroupNormBp(; kwargs...)

Describes a building process for a Groupnorm layer. make(channels, bp::CyGroupNorm)

Keyword arguments:

activation: activation function (default relu)

groups: the number of groups passed to GroupNorm constructor

init_shift: see initβ (default zeros32)

init_scale: see initγ (default ones32)

affine: see affine (default true)

track_stats: see track_stats (default false)

epsilon: see ϵ (default 1.0e-5)

momentum: see momentum (default 0.1)

Cyanotype.HybridAtrouConvBp — Type

aka Hybrid Dilated Convolution paper example example

Keyword arguments:

dilation_rates is not documented

convolution is not documented

Cyanotype.InstanceNormBp — Type

InstanceNormBp(; kwargs...)

Describes a building process for a InstanceNorm layer. make(channels, bp::CyInstanceNorm)

Keyword arguments:

activation: activation function (default relu)

init_shift: see initβ (default zeros32)

init_scale: see initγ (default ones32)

affine: see affine (default false)

track_stats: see track_stats (default false)

epsilon: see ϵ (default 1.0e-5)

momentum: see momentum (default 0.1)

Cyanotype.KwargsMapping — Type

KwargsMapping(; flux_function = :notflux, field_names = (), flux_kwargs = (),
                field_types = (), def_values = ())

Define a mapping of keyword arguments to interface a blueprint with a Flux function or constructor.

Cyanotype.NConvBp — Type

Template describing a module with N NConvBp repeated.

Keyword arguments:

convolution is not documented

nrepeat is not documented

Cyanotype.NoNormBp — Type

NoNormBp()

Keyword arguments:

activation: activation function (default Flux.relu)

Cyanotype.SqueezeExciteBp — Type

Keyword arguments:

volumetric: indicates a building process for three-dimensionnal data (default false)

activation: activation function (default Flux.relu)

gate_activation is not documented

reduction is not documented

Cyanotype.cyanotype — Function

cyanotype(bp::AbstractBlueprint; kwargs...)

Create a new blueprint from bp with the modifications defined by kwargs. This method is automatically generated by the @cyanotype macro during the process of defining a blueprint.

Cyanotype.make — Function

Cyanotype.spread — Method

spread(bp; kwargs...)

Return a new blueprint with kwargs spreaded over all fields of bp. This function allows to modify all nested cyanotypes at once.

# Nice printing of large objects
using GarishPrint

# Create a blueprint for a Hybrid A-trou Convolution module
hac = CyHybridAtrouConv();
# By default activation functions are relu
pprint(hac)

# Create a blueprint for a double convolution module with the second convolution as a usual
# convolutionnal layer
conv2 = CyDoubleConv(; convolution1 = hac, convolution2 = CyConv());
# By default activation functions are relu
pprint(conv2)

# Now let change all activation functions from relu to leakyrelu
spread(conv2; activation = leakyrelu) |> pprint

Cyanotype.@cyanotype — Macro

@cyanotype(doc, expr)
@cyanotype(doc, kmap, expr)

Defines a blueprint DataType with documentation doc and a struct declaration defined in expr. If the blueprint directly refers to a Flux function or constructor, kmap is the name of keyword arguments mapping. If there is some fields documentation in expr, it is automatically appended to doc.

Automatically generated functions:

FooBluePrint(; kwargs...): keyword argument constructor for FooBluePrint

mapping(::FooBluePrint): return, if defined, the mapping kmap

cyanotype(bp::FooBluePrint)

kwargs(bp::FooBluePrint): return, if defined, a Dict with the keyword arguments for

the Flux function or constructor, it can be used as follow: flux_function(arg1, arg2; kwargs(bp)...)

Example:

using Cyanotype: @cyanotype

@cyanotype (
"""
A FooBlueprint as example.
"""
) (
struct FooBlueprint{A<:Function}
    """`activation`: activation function"""
    activation::A = relu
end
)

For the keyword arguments mapping usage, see KwargsMapping documentation.